# src/main/resources/api/data-repository-openapi.yaml
openapi: 3.0.1
info:
  title: Data Repository API
  description: >
    This document defines the REST API for Data Repository.
    **Status: design in progress**

    There are four top-level endpoints (besides some used by swagger):
     * /swagger-ui.html - generated by swagger: swagger API page that provides this documentation and a live UI for
         submitting REST requests
     * /status - provides the operational status of the service
     * /api    - is the authenticated and authorized Data Repository API
     * /ga4gh/drs/v1 - is a transcription of the Data Repository Service API

    The overall API (/api) currently supports two interfaces:
     * Repository - a general and default interface for initial setup, managing ingest and repository metadata
     * Resource - an interface for managing billing accounts and resources

    The API endpoints are organized by interface. Each interface is separately versioned.

    ## Notes on Naming
    All of the reference items are suffixed with "Model". Those names are used as the class names in the generated Java
    code. It is helpful to distinguish these model classes from other related classes, like the DAO classes and the
    operation classes.

    ## Editing and debugging
    I have found it best to edit this file directly to make changes and then use the swagger-editor to validate. The
    errors out of swagger-codegen are not that helpful. In the swagger-editor, it gives you nice errors and links to
...

...
    ```
...

...

...

...

...

...
servers:
...
security:
  - googleoauth:
      - openid
      - email
      - profile
paths:
...
  /api/resources/v1/profiles:
    get:
      tags:
        - resources
      description: >
        Returns a list of all of the billing profiles
      operationId: enumerateProfiles
      parameters:
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
            set.
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: The numbers of items to return.
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: List of billing profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnumerateBillingProfileModel'
    post:
      tags:
        - resources
      description: >
        Creates a new profile associated with a billing account. This is asynchronous.
        The final return is of type definitions/BillingProfileModel.
      operationId: createProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingProfileRequestModel'
        required: false
      responses:
        200:
          description: Redirect for profile successful creation
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status of profile creation job & url for polling in the
            response header
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - see error message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to create profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      x-codegen-request-body-name: billingProfileRequest
  /api/resources/v1/profiles/{id}:
    get:
      tags:
        - resources
      description: Retrieve a billing profile by id
      operationId: retrieveProfile
      parameters:
        - name: id
          in: path
          description: A UUID to used to identify an object in the repository
          required: true
          schema:
            type: string
      responses:
        200:
          description: Billing profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingProfileModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see billing profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - billing profile id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
    delete:
      tags:
        - resources
      description: >
        Delete a billing profile by id. This is asynchronous.
        The final return is of type definitions/DeleteResponseModel.
      operationId: deleteProfile
      parameters:
        - name: id
          in: path
          description: A UUID to used to identify an object in the repository
          required: true
          schema:
            type: string
      responses:
        200:
          description: Redirect for profile delete result
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status of profile delete job & url for polling in the response
            header
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to delete profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - profile does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        409:
          description: Profile in use by at least one dataset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      put:
        tags:
          - resources
        description: >
          Update a billing profile by id. This is asynchronous.
          The final return is of type definitions/UpdateResponseModel.
        operationId: updateProfile
        parameters:
          - name: id
            in: path
            description: A UUID to used to identify an object in the repository
            required: true
            schema:
              type: string
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingProfileRequestModel'
          required: true
        responses:
          200:
            description: Redirect for profile update result
            headers:
              location:
                description: url for the job result
                schema:
                  type: string
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JobModel'
          202:
            description: Job status of profile update job & url for polling in the response
              header
            headers:
              location:
                description: url for the job polling
                schema:
                  type: string
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JobModel'
          400:
            description: Bad request - invalid id, badly formed
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorModel'
          403:
            description: No permission to update profile
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorModel'
          404:
            description: Not found - profile does not exist
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorModel'
          409:
            description: Profile in use by at least one dataset
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorModel'
  '/api/resources/v1/profiles/{id}/policies/{policyName}/members':
    post:
      tags:
        - resources
      description: Adds a member to the specified policy for the snapshot
      operationId: addProfilePolicyMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyMemberRequest'
        required: true
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/ProfilePolicyName'
      responses:
        201:
          description: Policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        400:
          description: Bad request - invalid id, badly formed, or email not found
          schema:
            $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to update policies
          schema:
            $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - profile id does not exist
          schema:
            $ref: '#/components/schemas/ErrorModel'
...
  /api/repository/v1/jobs/{id}:
    get:
      tags:
        - repository
      description: Retrieve a job's status by id
      operationId: retrieveJob
      parameters:
        - name: id
          in: path
          description: A UUID to used to identify an object in the repository
          required: true
          schema:
            type: string
      responses:
        200:
          description: Redirect for object successful creation
          headers:
            location:
              description: url for the job result
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        202:
          description: Job status
          headers:
            location:
              description: url for the job polling
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobModel'
        400:
          description: Bad request - invalid id, badly formed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        403:
          description: No permission to see job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        404:
          description: Not found - job id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
    delete:
      tags:
        - repository
      description: Delete the job and data associated with it
      operationId: deleteJob
      parameters:
        - name: id
          in: path
          description: A UUID to used to identify an object in the repository
          required: true
          schema:
            type: string
      responses:
        204:
          description: Job was deleted
          content: {}
...
components:
  schemas:
...
    UniqueIdProperty:
      pattern: ^[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}
      type: string
      description: >
        Unique identifier for a dataset, snapshot, etc.
    BillingProfileRequestModel:
      required:
        - biller
        - billingAccountId
        - id
        - profileName
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UniqueIdProperty'
        billingAccountId:
          type: string
          description: unique identifier of the billing account from Google
        profileName:
          type: string
          description: display name to be used for the profile;
        biller:
          type: string
          description: either “direct” or the reseller of the sub-account that sends
            the bill (like “on_spend”)
        description:
          type: string
          description: Free text that describes the profile
    BillingProfileModel:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UniqueIdProperty'
        billingAccountId:
          type: string
          description: unique identifier of the billing account from Google
        profileName:
          type: string
          description: display name to be used for the profile;
        biller:
          type: string
          description: either “direct” or the reseller of the sub-account that sends
            the bill (like “on_spend”)
        description:
          type: string
          description: Free text that describes the profile
        createdDate:
          type: string
          description: Date the profile was created
        createdBy:
          type: string
          description: Identity of the user who originally created the profile
    EnumerateBillingProfileModel:
      type: object
      properties:
        total:
          type: integer
          description: Total number of billing profiles
        items:
          type: array
          items:
            $ref: '#/components/schemas/BillingProfileModel'
      description: >
        The total number of billing profiles available and a page of profiles
...
    JobModel:
      required:
        - id
        - job_status
        - status_code
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UniqueIdProperty'
        description:
          type: string
          description: Description of the job's flight from description flight input
            parameter
        job_status:
          type: string
          description: Status of job
          enum:
            - running
            - succeeded
            - failed
        status_code:
          type: integer
          description: HTTP code
        submitted:
          type: string
          description: Timestamp when the flight was created
        completed:
          type: string
          description: Timestamp when the flight was completed; not present if not
            complete
      description: >
        Status of job
    ErrorModel:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        errorDetail:
          type: array
          items:
            type: string
      description: >
        Representation of an error that occurred while processing a request.
        Some errors provide error details. For example, validation of input may report several problems.
        Each of those are reported in a separate error detail.
...
    PolicyMemberRequest:
      required:
        - email
      type: object
      properties:
        email:
          type: string
      description: >
        email of user or group to add to policy
    PolicyModel:
      type: object
      properties:
        name:
          type: string
        members:
          type: array
          items:
            type: string
      description: >
        Describes a policy
    PolicyResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicyModel'
      description: >
        email of user or group to add to policy
...
  parameters:
    Id:
      name: id
      in: path
      description: A UUID to used to identify an object in the repository
      required: true
      schema:
        type: string
...
    ProfilePolicyName:
      name: policyName
      in: path
      description: The relevant policy
      required: true
      schema:
        type: string
        enum:
          - owner
          - user
...
  securitySchemes:
    googleoauth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
    authToken:
      type: apiKey
      description: >
        A valid authorization token must be passed in the 'Authorization' header.
        Example syntax for using 'Authorization' header :
          Bearer: token_string
      name: Authorization
      in: header